requirements.txt
google-cloud-storage==1.42.0
cloudpathlib==0.6.0
absl-py==0.12.0
pathos>=0.2.7
pyfarmhash==0.2.2
matplotlib>=3.0.3
numpy==1.20.2
pyDOE==0.3.8
scipy==1.6.2
seaborn==0.9.0
lxml==4.5.2
cvxopt==1.2.6
cvxpy==1.1.12
dp-accounting==0.0.2
tqdm>=4.47.0
fsspec==2021.7.0
gcsfs==2021.7.0
pandas>=1.2.5
git+git://github.com/world-federation-of-advertisers/cardinality_estimation_evaluation_framework/@master#egg=wfa_cardinality_estimation_evaluation_framework
